refactor/ffi: small ffi refactors
