+++ src/ffi/mod.rs
@@ -54,7 +54,8 @@ use core::client::Client;
 use core::errors::CoreError;
 use core::translated_events::NetworkEvent;
 use ffi::errors::FfiError;
-use ffi::nfs::get_file_writer::{FfiWriterHandle, GetFileWriter};
+use ffi::nfs::FfiWriterHandle;
+use ffi::nfs::get_file_writer::GetFileWriter;
 use ffi::nfs::create_file::CreateFile;
 use libc::{c_char, int32_t, int64_t};
 use maidsafe_utilities::log as safe_log;
